Sushi_5C.bedpe {Sushi}	R Documentation

Sushi_5C.bedpe

Description

This data set list the genomic locations of 5C interactions in multiple cell lines with coordinates based on the NCBI36 / hg18 genome build.

Usage

Sushi_5C.bedpe

Format

bedpe format

Source

Sanyal, A., Lajoie, B. R., Jain, G. & Dekker, J. The long-range interaction landscape of gene promoters. Nature 489, 109-113 (2012).

[Package Sushi version 1.20.0 Index]