Sushi_ChIAPET_pol2.bedpe {Sushi}	R Documentation

Sushi_ChIAPET_pol2.bedpe

Description

This data set list the genomic locations of Pol2 ChIA PET interactions in K562 cells with coordinates based on the NCBI36 / hg18 genome build.

Usage

Sushi_ChIAPET_pol2.bedpe

Format

bedpe format

Source

Li, G. et al. Extensive Promoter-Centered Chromatin Interactions Provide a Topological Basis for Transcription Regulation. Cell 148, 84-98 (2012).

[Package Sushi version 1.20.0 Index]