Sushi_DNaseI.bedgraph {Sushi}	R Documentation

Sushi_DNaseI.bedgraph

Description

This data set describes read depths across the genomce resulting from a DNaseI hypersensitivity experiment in K562 cells with coordinates based on the NCBI36 / hg18 genome build.

Usage

Sushi_DNaseI.bedgraph

Format

bedgraph format

Source

Neph, S. et al. An expansive human regulatory lexicon encoded in transcription factor footprints. Nature 489, 83-90 (2012).

[Package Sushi version 1.20.0 Index]