Sushi_GWAS.bed {Sushi}	R Documentation

Sushi_GWAS.bed

Description

Bed data representing results from a GWAS study of blood pressure and cardiovascular disease risk with coordinates based on the NCBI36 / hg18 genome build.

Usage

Sushi_GWAS.bed

Format

bed format

Source

Ehret, G. B. et al. Genetic variants in novel pathways influence blood pressure and cardiovascular disease risk. Nature 478, 103-109 (2011).

[Package Sushi version 1.20.0 Index]