Sushi_HiC.matrix {Sushi}	R Documentation

Sushi_HiC.matrix

Description

Bed data representing results from a GWAS study of blood pressure and cardiovascular disease risk with coordinates based on the NCBI36 / hg18 genome build.

Usage

Sushi_HiC.matrix

Format

matrix

Source

Dixon, J. R. et al. Topological domains in mammalian genomes identified by analysis of chromatin interactions. Nature (2012). doi:10.1038/nature11082

[Package Sushi version 1.20.0 Index]