labelgenome {Sushi}	R Documentation

Adds genome coordinates to the x-axis of a Sushi plot

Description

Adds genome coordinates to the x-axis of a Sushi plot

Usage

labelgenome(chrom, chromstart, chromend, genome = NULL, space = 0.01,
  scale = "bp", side = 1, scipen = 20, n = 5, chromfont = 2,
  chromadjust = 0.015, chromcex = 1, chromline = 0.5, scalefont = 2,
  scaleadjust = 0.985, scalecex = 1, scaleline = 0.5, line = 0.18,
  edgeblankfraction = 0.1, ...)

Arguments

chrom
chromosome to plot

chromstart
start position

chromend
end position

genome
a genome object (2 columns: column 1 = chromosome name, column 2 = length of chromosome). Only for multi chromosomal plots

space
the space in between each chromosome as a fraction of the width of the plot. Only for multi chromosomal plots

scale
Scale of the plot ('bp','Kb','Mb')

side
Side of the scale to add the plot to. Only tested for sides 1 and 3.

scipen
higher values decrease the likelihood of using scientific for the position labels.

n
Desired number of ticks

chromfont
font type of chromosome label

chromadjust
position, as a fraction of the width of the plot, of the chomosome label

chromcex
font size of the chomosome label

chromline
vertical offset of the chomosome label

scalefont
font type of scale label

scaleadjust
position, as a fraction of the width of the plot, of the scale label

scalecex
font size of the scale label

scaleline
vertical offset of the scale label

line
vertical offset of position labels

edgeblankfraction
percent of the edges to leave black for chromosome and scale labels

...
values to be passed to axis

Examples

data(Sushi_DNaseI.bedgraph)
# set the genomic regions

plotBedgraph(Sushi_DNaseI.bedgraph,chrom="chr11",chromstart=1650000,chromend=2350000,colorbycol=SushiColors(7))
labelgenome(chrom="chr11",chromstart=1650000,chromend=2350000,side=1,n=4,scale="Mb")
axis(side=2,las=2,tcl=.2)
mtext("Read Depth",side=2,line=1.75,cex=.75,font=2)

[Package Sushi version 1.20.0 Index]