labelplot {Sushi}	R Documentation

adds a letter and a title to a plot

Description

This function adds a letter and a title (both are optional) to the top of a plot. Udeful for generating paper figures.

Usage

labelplot(letter = NULL, title = NULL, letteradj = -0.05, titleadj = 0,
  letterfont = 2, titlefont = 2, lettercex = 1.2, titlecex = 1,
  letterline = 0.5, titleline = 0.5, lettercol = "black",
  titlecol = "black")

Arguments

letter
A string, typically a letter or number (eg 'A', 'A)', '1', etc) to lable the plot with

title
A string for a plot title

letteradj
adj of letter. See par

titleadj
adj of title. See par

letterfont
font of letter. See par

titlefont
font of title See par

lettercex
cex of letter. See par

titlecex
cex of title See par

letterline
line of letter. See par

titleline
line of title See par

lettercol
color of letter. See par

titlecol
color of title See par

Examples

par(mar=c(3,3,3,3))
plot((1:10),col=maptocolors(vec=(1:10),colorRampPalette(c("blue","red"))),pch=19,cex=4)
labelplot("A)"," sample plot",lettercex=2,titlecex=2,titlecol="blue")

[Package Sushi version 1.20.0 Index]