plotHic {Sushi}	R Documentation

plots HiC interactio matrix

Description

plots HiC interactio matrix

Usage

plotHic(hicdata, chrom, chromstart, chromend, max_y = 30, zrange = NULL,
  palette = SushiColors(7), flip = FALSE)

Arguments

hicdata
interaction matrix representing HiC data. Row and column names should be postions along a chromosome

chrom
chromosome of region to be plotted

chromstart
start position

chromend
end position

max_y
The maximum bin distance to plot

zrange
The range of interaction scores to plot (more extreme value will be set to the max or min)

palette
color palette to use for representing interaction scores

flip
TRUE/FALSE whether plot should be flipped over the x-axis

Examples

data(Sushi_HiC.matrix)

chrom            = "chr11"
chromstart       = 500000
chromend         = 5050000

phic = plotHic(Sushi_HiC.matrix,chrom,chromstart,chromend,max_y = 20,zrange=c(0,28),palette = topo.colors,flip=FALSE)

labelgenome(chrom,chromstart,chromend,side=1,scipen=20,n=4,scale="Mb",edgeblankfraction=0.20,line=.18,chromline=.5,scaleline=0.5)

addlegend(phic[[1]],palette=phic[[2]],title="score",side="right",bottominset=0.4,topinset=0,xoffset=-.035,labelside="left",width=0.025,title.offset=0.035)

[Package Sushi version 1.20.0 Index]